Upgrade Errorprone
Upgrade errorprone to 2.13.1.  Like all errorprone upgrades there are
new checks requiring code changes.

* Unused methods now cause compilation errors
* fields must be static and final
* Effectively constant booleans must now be returned as true/false.
* longs should not auto-cast to double.

+++ metrics/core/src/main/java/org/hyperledger/besu/metrics/opentelemetry/OpenTelemetrySystem.java
@@ -277,7 +277,7 @@ public class OpenTelemetrySystem implements ObservableMetricsSystem {
             resultLongObserver -> {
               for (int i = 0; i < garbageCollectors.size(); i++) {
                 resultLongObserver.observe(
-                    garbageCollectors.get(i).getCollectionTime(), labelSets.get(i));
+                    (double) garbageCollectors.get(i).getCollectionTime(), labelSets.get(i));
               }
             });
     final AttributeKey<String> typeKey = AttributeKey.stringKey(TYPE_LABEL_KEY);
